requirements.txt
alembic==1.13.1
Flask==3.1.1
flask_cors==5.0.1
Flask_JWT_Extended==4.7.1
Flask_Login==0.6.3
Flask_Migrate==3.1.0
flask_sqlalchemy==3.1.1
grpcio==1.71.0
protobuf==6.31.0
SQLAlchemy==2.0.23
transliterate==1.10.2
Werkzeug==3.1.3
psycopg2-binary==2.9.10
Pillow==10.1.0